updated delcontainer

+++ dockerdelcontainers.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 RMLIST=$(sudo docker ps -a -q)
-sudo docker stop $RMLIST
-sudo docker rm $RMLIST 
+if [ $RMLIST == "" ] then;
+	echo "未找到运行中的docker容器"
+else
+	echo "停止并删除运行中的容器" &&\
+	sudo docker stop $RMLIST &&\
+	sudo docker rm $RMLIST
+fi
